Serve loan index from Metabackend.
This is the first iteration of this functionality.

We can currently index a loan, and then query this index by different
fields. Additionally, blockchain constants (loan creation parameters are
ceched).

We do not return loan state, whether this functionality should be
implemented in the metabackend is to be decided.

To do:
 - Tests! We currently don't even have unit tests.
 - Versioning of contracts, but this requires some design.
 - Verification of contract bytecode when indexing. Currently we let
 anything be indexed as a loan as long as it responds to the loan ABI.
 In the future, we will need to store information on whether a given
 loan corresponds to a version of the loan contract that we know.