Refactor getline.ts API
This is probably still not final, but we're getting there.

We've introduced a few more types to clean up access to tokens,
addresses and balances. We've also made web3 types opaque and wrapped
them in the Contract type.

Finally, we also added a lock on the retrieval of contracts from the
metabackend to make sure we don't send out too many requests at once.