Add interface to print tokens from getline.ts
This creates a new PrintableToken class and functions to convert an
ordinary token into a printable one. The main testToken is also switched
to be a PrintableToken.

We leave the main Loan class as emitting plain Tokens for now, since we
want to limit the leak of demo code to a minimum - and Client.testToken
is already part of that.

We will have to also consider adding contract typechecking in the client
- either directly on the blockchain or employ the metabackend to help us
out in this. This is also related to server-side verification of loans
that are indexed, but I think this is out of scope for the demo. Let's
do a proper design for this some day.