Make mutating blockchain functions always block.
We introduce a new Contract method: .mutate. This is similar to .call,
except it's made for state mutating smart contract methods. Since these
methods generate a transaction, we have to wait until this transaction
gets mined until we can see the results reflected.

We also updated the Loan and Token methods that mutate state to use this
new function. Consumers of the API are of course free to not block on
the call if they don't mind stale reads (or unconfirmed mutations).