Switch to asynchronous eth.network and eth.account
This means that we can no longer initialize the entire getline.ts client
library in the constructor. We add an async initialize() method for
explicit initializations, and a crutch to call it from existing methods
if the client hasn't yet been explicitely initialized.

Adding this method is actually quite serendipitous, as it will let us
have a single point where clients can call it to get exceptions thrown
if something is wrong with the blockchain we're connected to (no web3
provider, no account unlocked, wrong network, etc). We'll have to add a
separate class of errors for that.