Replace `yarn generate; yarn tsc` with a unified `yarn build`
This script will:
 - generate proto definitions
 - run tsc
 - copy over generated .js/.d.ts from src/generated to dist/generated

The last step seems to be necessary after disabling allowJs (which we
want because we want to generate declaration files). Typescript doesn't
seem to treat typed JS as a first-class citizen enough to emit it on
its' own.

This should still be replaced at some point with a real build system
that understands our monorepo layout (Bazel?).