Add .loansByState method to getline.ts
We also up the tsc lib to es2016. This in turn made some tests unstable
(because of array ordering?), so I reworked some of them.

The current metabackend implementation for this always polls the
blockchain for status - this will obviously have to be either cached in
memory or in the database at some point, but it's enough to ship a
working implementation for the client to use.