Fix bignumber.js versioning discrepencies
This was noticed in a debugging session by balinskia & q3k.

 - The Frontend creates a BigNumber (from its' own module) and passes it
   to a getline.ts function.
 - Getline.ts takes this BigNumber and passes it as an argument to a
   web3.js contract call.
 - web3.js then checks arguments passed and removes them if they're
   objects but not BigNumbers. Our BigNum is removed because both of the
   following checks fail:
     - `n instanceof BigNumber` -> fails because the passed BigNumber
       instance is not an instance of web3's BigNumber constructor
     - `n.constructor.name === "BigNumber"` -> fails because the passed
       BigNumber has its' class name mangled by UglifyJs

There's two ways to fix this issue:

 - Make sure that web3.js and the frontend use the same version of the
   BigNumber constructor.
 - Stop mangling names in UglifyJs.

We opt for the first option, as disabling mangling grows our app bundle
by around 15%. To implement this, we make getline.ts export the BigNumber
constructor to clients. We then make the frontend use this constructor.

Additionally, we bump web3.js to use the upstream BigNubmer library
instead of their own fork of it. As far as I understand, this fork
exists for the sole reason of compatibility with running web3.js under
the Otto VM (which we don't plan on supporting).