Redeploy smart contracts
The previous fix was in InvestorLedger, which is not a leaf contract.
Thus we had to perform a redeploy of everything (we can't redelploy just
affected leaves just yet...).

We also cleaned up the migrations script to not deploy useless
contracts, and then made metabackend serve contracts that are not fully
deployed (like the Loan or any of the *Token) contracts.

We also redeploy the metabackend.