Do not re-trigger GET_MY_INVESTMENTS_ACTION
This is already done when MyInvestmentTiles appear, and are conditional
to their appearance (they only show up if !state.isLoading, and that is
is set by the mutation RECEIVED_MY_INVESTMENTS which is triggered by
this action). This causes an infinite loop to happen when we try to load
this component.